Web Inspector: Audits: don't cache default audits https://bugs.webkit.org/show_bug.cgi?id=192918 <rdar://problem/46626543>
Reviewed by Brian Burg.

Instead of adding the default audits to the "audits" `WI.ObjectStore`, which preserves them
across WebInspector sessions, load them every time WebInspector is opened (unless there are
existing audits) so that the localized strings are able to respond to locale changes.

* UserInterface/Controllers/AuditManager.js:
(WI.AuditManager.prototype.removeTest):
(WI.AuditManager.prototype.addDefaultTestsIfNeeded):


Canonical link: https://commits.webkit.org/207490@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@239449 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebInspectorUI/UserInterface/Controllers/AuditManager.js
@@ -172,7 +172,8 @@ WI.AuditManager = class AuditManager extends WI.Object
 
         this.dispatchEventToListeners(WI.AuditManager.Event.TestRemoved, {test});
 
-        WI.objectStores.audits.deleteObject(test);
+        if (!test.__default)
+            WI.objectStores.audits.deleteObject(test);
     }
 
     // Private
@@ -253,8 +254,8 @@ WI.AuditManager = class AuditManager extends WI.Object
         ];
 
         for (let test of defaultTests) {
+            test.__default = true;
             this._addTest(test);
-            WI.objectStores.audits.addObject(test);
         }
     }
 };
